Minor refactor of rdp_out_unistr()
The logon packets wants a mandatory null termination
to be written to stream if not string is available.
The current design was to do if statements at those
points to handle this. Refactored that code into
function rdp_out_unistr_mandatory_null() for a
cleaner code when building up the packets.