General network fixes
Major change in this commit is to resolve address given via command
line as early as possible and only once during an instance lifetime
of rdesktop. This change fixes the problems with network error or
pending resize reconnect (Windows 2008) which could resolve to another
IP if DNS is configured for round robin and with that, network connect
cookie is invalidated.

I also took the liberty to remove IPv6 from being a compile time options
and make it always used.

This change also fixed problems with CredSSP where a FQDN was forced to
be entered as servername via command line to work as expected. Now it would
use the getnameinfo() to get FQDN for building a complete service name.