Set master version to 1.8.99 to be greater than 1.8.x versions
Currently, the version number of releases made from the branch 1.8.x are
greater than the version in the master branch (1.8.3post).

That makes life a little harder for packages wanting to package the code
from the master branch.

So, set the version to 1.8.99, ensuring with a high probability, that
the version will be greater than all 1.8.x releases.

Fixes: https://github.com/rdesktop/rdesktop/issues/325